Category screen header alignment

+++ BeautyPop/BeautyPop/FeedViewAdapter.swift
@@ -93,7 +93,7 @@ class FeedViewAdapter {
     }
     
     static func getFeedItemCellSize(width: CGFloat) -> CGSize {
-        let availableWidthForCells: CGFloat = width - (Constants.FEED_ITEM_SIDE_SPACING * (Constants.FEED_COLUMNS + 1))  // left middle right spacing
+        let availableWidthForCells: CGFloat = width - (Constants.FEED_ITEM_SIDE_SPACING * CGFloat(Constants.FEED_COLUMNS + 1))  // left middle right spacing
         let cellWidth: CGFloat = availableWidthForCells / Constants.FEED_COLUMNS
         let cellHeight = cellWidth + Constants.FEED_ITEM_DETAILS_HEIGHT
         return CGSizeMake(cellWidth, cellHeight)
